Fix all the JWT issues & refactorings
