+++ src/TrimmedTodo.MinimalApi.Dapper.Sqlite/TrimmedTodo.MinimalApi.Dapper.Sqlite.csproj
@@ -11,6 +11,14 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Include="..\HelloWorld.Web\WebApplicationExtensions.cs" Link="WebApplicationExtensions.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Include="JwtConfigHelper.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <!-- Manually marking these assemblies as trimmable until we update them -->
     <TrimmableAssembly Include="Microsoft.IdentityModel.Tokens" />
     <TrimmableAssembly Include="Microsoft.AspNetCore.Components.Server" />
