+++ src/TrimmedTodo.WebApi.EfCore.Sqlite/Program.cs
@@ -1,8 +1,5 @@
-using System.Text.Json;
-using System.Text.Json.Serialization;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.IdentityModel.Tokens;
 using TrimmedTodo.WebApi.EfCore.Sqlite.Data;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -13,26 +10,12 @@ if (!builder.Environment.IsDevelopment())
 }
 
 builder.Services.AddAuthentication()
-    .AddJwtBearer(o =>
-    {
-        if (!builder.Environment.IsDevelopment())
-        {
-            // When not running in development configure the JWT signing key from environment variable
-            var jwtKeyMaterialValue = builder.Configuration["JWT_SIGNING_KEY"];
-
-            if (string.IsNullOrEmpty(jwtKeyMaterialValue))
-                throw new InvalidOperationException("JWT signing key not found!");
-
-            var jwtKeyMaterial = Convert.FromBase64String(jwtKeyMaterialValue);
-            var jwtSigningKey = new SymmetricSecurityKey(jwtKeyMaterial);
-            o.TokenValidationParameters.IssuerSigningKey = jwtSigningKey;
-        }
-    });
+    .AddJwtBearer(JwtConfigHelper.ConfigureJwtBearer(builder));
 
 builder.Services.AddAuthorization();
 
 builder.Services.AddOptions<JwtBearerOptions>(JwtBearerDefaults.AuthenticationScheme)
-    .Validate<IHostEnvironment, ILoggerFactory>(ValidateJwtOptions,
+    .Validate<IHostEnvironment, ILoggerFactory>(JwtConfigHelper.ValidateJwtOptions,
         "JWT options are not configured. Run 'dotnet user-jwts create' in project directory to configure JWT.")
     .ValidateOnStart();
 
@@ -55,38 +38,7 @@ app.UseSwaggerUI();
 
 app.MapControllers();
 
-await app.StartAsync();
-
-if (builder.Configuration["SHUTDOWN_ON_START"] != "true")
-{
-    app.WaitForShutdown();
-}
-else
-{
-    await app.StopAsync();
-}
-
-static bool ValidateJwtOptions(JwtBearerOptions options, IHostEnvironment hostEnvironment, ILoggerFactory loggerFactory)
-{
-    var relevantOptions = new JwtOptionsSummary
-    {
-        Audience = options.Audience,
-        ClaimsIssuer = options.ClaimsIssuer,
-        Audiences = options.TokenValidationParameters?.ValidAudiences,
-        Issuers = options.TokenValidationParameters?.ValidIssuers,
-        IssuerSigningKey = options.TokenValidationParameters?.IssuerSigningKey?.ToString()
-    };
-    if ((string.IsNullOrEmpty(relevantOptions.Audience) && relevantOptions.Audiences?.Any() != true)
-        || (relevantOptions.ClaimsIssuer is null && relevantOptions.Issuers?.Any() != true)
-        || string.IsNullOrEmpty(relevantOptions.IssuerSigningKey))
-    {
-        return false;
-    }
-    var logger = loggerFactory.CreateLogger(hostEnvironment.ApplicationName ?? nameof(Program));
-    logger.LogInformation("JwtBearerAuthentication options configuration: {JwtOptions}",
-        JsonSerializer.Serialize(relevantOptions, ProgramJsonSerializerContext.Default.JwtOptionsSummary));
-    return true;
-}
+await app.StartApp("/api/todos");
 
 static async Task EnsureDb(string cs, IServiceProvider services, ILogger logger)
 {
@@ -94,18 +46,4 @@ static async Task EnsureDb(string cs, IServiceProvider services, ILogger logger)
 
     using var db = services.CreateScope().ServiceProvider.GetRequiredService<TodoDb>();
     await db.Database.MigrateAsync();
-}
-
-internal class JwtOptionsSummary
-{
-    public string? Audience { get; set; }
-    public string? ClaimsIssuer { get; set; }
-    public IEnumerable<string>? Audiences { get; set; }
-    public IEnumerable<string>? Issuers { get; set; }
-    public string? IssuerSigningKey { get; set; }
-}
-
-[JsonSerializable(typeof(JwtOptionsSummary))]
-internal partial class ProgramJsonSerializerContext : JsonSerializerContext
-{
 }
